+++ packages/atomic/cypress/utils/componentUtils.ts
@@ -16,10 +16,10 @@ export function doSortOccurences(originalValues: string[]) {
     );
 }
 
-export const aliasNoAtSignBuilder = (alaisWithAtSign: IAlias) => {
-  const alaisNoAtSign = Object.assign({}, alaisWithAtSign);
-  Object.keys(alaisNoAtSign).forEach((key: string) => {
-    alaisNoAtSign[key] = alaisNoAtSign[key].split('@')[1];
+export const aliasNoAtSignBuilder = (aliasWithAtSign: IAlias) => {
+  const aliasNoAtSign = Object.assign({}, aliasWithAtSign);
+  Object.keys(aliasNoAtSign).forEach((key: string) => {
+    aliasNoAtSign[key] = aliasNoAtSign[key].split('@')[1];
   });
-  return alaisNoAtSign;
+  return aliasNoAtSign;
 };