+++ packages/headless/src/test/mock-thunk-extra-arguments.ts
@@ -1,0 +1,14 @@
+import {buildLogger} from '../app/logger';
+import {ThunkExtraArguments} from '../app/thunk-extra-arguments';
+
+export function buildMockThunkExtraArguments(
+  config: Partial<ThunkExtraArguments> = {}
+): ThunkExtraArguments {
+  return {
+    analyticsClientMiddleware: jest.fn(),
+    preprocessRequest: jest.fn(),
+    validatePayload: jest.fn(),
+    logger: buildLogger({level: 'silent'}),
+    ...config,
+  };
+}